Set zone heating schedule based on trv schedule (#9)
* Initial check in

Put framework in place to report to HiveHome if any TRVs heating schedule has changed. This will be used to flag when to set the zones heating schedule based on the sum of the child TRVs heating schedule.
Also more tidying up of the autoAlias attribute.

* More AutoAlias work

Only update a products alias if it has changed.

* More work on zone heading schedule

Refactor the functionality to deal with heating zones.
Get all TRVs in the heating zone and get their schedule and convert that into a JSON object.
TODO: merge the schedules into a new composite schedule.

* Add child lock

Add child lock

* Initial check in

Put framework in place to report to HiveHome if any TRVs heating schedule has changed. This will be used to flag when to set the zones heating schedule based on the sum of the child TRVs heating schedule.
Also more tidying up of the autoAlias attribute.

* More AutoAlias work

Only update a products alias if it has changed.

* More work on zone heading schedule

Refactor the functionality to deal with heating zones.
Get all TRVs in the heating zone and get their schedule and convert that into a JSON object.
TODO: merge the schedules into a new composite schedule.

* Add child lock

Add child lock

* Add logic to merge two day elements

Start work on adding logic to merge schedules.
Completed work on merging schedules. Need to finish work on reducing the number of day events if there are more than the maximum allowed.

* Fix new implementation bugs

Fix bugs and test the generated schedule.
It's not good, need to think how to create new logic to generate schedule.

* Updated schedule combining logic

Update the logic to combine each TRVs day schedule into the heating day schedule to consider the the heating elements before and after the TRV element.
Fix some perlisms.
Add some help to the product.

* Fix bug in autoAlias

autoAlias was using deviceType instead of productType. Only TRVs have this internal value.

* Test schedule update

Test to see if the combined TRV schedules if they match the heating schedule. Log the details.
Add a new attribute for heating products which is used to enable this new feature.

* Streamline the schedule process 

Change the process so that it only merges unique TRV schedules for the current day.
Only include TRV schedules that have the setScheduleFromTRV attribute set to 1.
Add setScheduleFromTRV to trvcontrol devices defaulted to 0.

* Fix a bug 

Fix a bug in generating the new heating schedule.
Remove some comments and a TODO.

* Add testing 

Add testing

* Correct schedule merging logic

* Fix missing logic

Missing logic to determine if the previous element added was of the same type to catch error if element temperature is higher than previous other element type temp.

* Simplify logic

Simplify the logic and remove unnecessary condition.
Remove code duplication.

* Perl copy hash ref

Work around Perls default to copy variables by reference which causes a variable to be incorrectly modified.

* Offline device

Stop HiveHome_Write_Product processing commands of offline products.

* Add functionality to reduce number of elements 

If after merging TRV heating schedules the number of elements are more than allowed by the interface. Repeatedly merge elements by smallest temperature difference until the number of elements are valid.

* Enable the feature

Enable the feature so that the actual schedule is applied to the thermostat.