Initial work
Created a new library to hold common HiveHome functions. Moved some of the existing common functions to the new library.
Modified how the heating schedule and target are displayed so that they do not show the temperature offset. The temperature offset is applied to the schedule before it is sent to the device.
Modified how the target is used within the library to ensure the offset is applied to it where required.