More fixes
Renamed hhc_IsValidNumber to hhc_IsValidDuration as it is used for checking if the provided number of mins is valid for boost.
Added temperatureOffset parameter to HiveHome_ConvertUIDayProfileStringToCmdString so that the when the week profile shown on the UI is converted to a command script, the temp offset can be added.
Created a new hhc_IsValidTemperatureOffset function for verifying temp offsets separate to hhc_IsValidTemperature.
Changed some of the formatting around '{'.