Merge changes into master (#11)
* Initial work

Created a new library to hold common HiveHome functions. Moved some of the existing common functions to the new library.
Modified how the heating schedule and target are displayed so that they do not show the temperature offset. The temperature offset is applied to the schedule before it is sent to the device.
Modified how the target is used within the library to ensure the offset is applied to it where required.

* Fix errors in the offsetTemperatuire functionality

Fix errors in the offset temperature helper functions. Fix errors in the IsValidTemperature helper function. Add forceUpdateSchedule attribute to support testing of this feature.
When the offsetTemperature attribute is modified, force the entire week schedule to be pushed to the heating device to ensure the new offset is applied.

* More fixes

Renamed hhc_IsValidNumber to hhc_IsValidDuration as it is used for checking if the provided number of mins is valid for boost.
Added temperatureOffset parameter to HiveHome_ConvertUIDayProfileStringToCmdString so that the when the week profile shown on the UI is converted to a command script, the temp offset can be added.
Created a new hhc_IsValidTemperatureOffset function for verifying temp offsets separate to hhc_IsValidTemperature.
Changed some of the formatting around '{'.

* Work in progress

Fix some error formatting issues and an attribute definition issue.
Correct logic in setting the temperatureOffset attribute, was using the full day name rather than the short day name.

* Work around 

Allow the set attribute temperatureOffset to set the schedule and disable the automatic offsetting of the temperature which would use the existing offset rather than the new offset. The temperature is offset before the call so that the correct schedule is applied.

* WIP

Fix errors in functions which try to use undefined variables.
Modified set attribute for temperatureOffset so that it can only update the week profile when the device is not disconnected and it is initialised.
Offset temperature for all days schedule while parsing.
Reduce the logging levels.

* Revert heating to schedule

If the number of TRVs requesting heat is less than the required amount, revert the heating to schedule from scheduleOverride. Originally this required no TRVs to be calling for heat for this to work.

* Fix errors in the offsetTemperatuire functionality

Fix errors in the offset temperature helper functions. Fix errors in the IsValidTemperature helper function. Add forceUpdateSchedule attribute to support testing of this feature.
When the offsetTemperature attribute is modified, force the entire week schedule to be pushed to the heating device to ensure the new offset is applied.

* Work in progress

Fix some error formatting issues and an attribute definition issue.
Correct logic in setting the temperatureOffset attribute, was using the full day name rather than the short day name.

* WIP

Fix errors in functions which try to use undefined variables.
Modified set attribute for temperatureOffset so that it can only update the week profile when the device is not disconnected and it is initialised.
Offset temperature for all days schedule while parsing.
Reduce the logging levels.

* Fix errors in the offsetTemperatuire functionality

Fix errors in the offset temperature helper functions. Fix errors in the IsValidTemperature helper function. Add forceUpdateSchedule attribute to support testing of this feature.
When the offsetTemperature attribute is modified, force the entire week schedule to be pushed to the heating device to ensure the new offset is applied.

* Work in progress

Fix some error formatting issues and an attribute definition issue.
Correct logic in setting the temperatureOffset attribute, was using the full day name rather than the short day name.

* WIP

Fix errors in functions which try to use undefined variables.
Modified set attribute for temperatureOffset so that it can only update the week profile when the device is not disconnected and it is initialised.
Offset temperature for all days schedule while parsing.
Reduce the logging levels.

* Fix errors in the offsetTemperatuire functionality

Fix errors in the offset temperature helper functions. Fix errors in the IsValidTemperature helper function. Add forceUpdateSchedule attribute to support testing of this feature.
When the offsetTemperature attribute is modified, force the entire week schedule to be pushed to the heating device to ensure the new offset is applied.

* Work in progress

Fix some error formatting issues and an attribute definition issue.
Correct logic in setting the temperatureOffset attribute, was using the full day name rather than the short day name.

* WIP

Fix errors in functions which try to use undefined variables.
Modified set attribute for temperatureOffset so that it can only update the week profile when the device is not disconnected and it is initialised.
Offset temperature for all days schedule while parsing.
Reduce the logging levels.

* c

c

* TRVs no longer have zone IDs

Remove the zone ids checks for TRVS, assume they all belong to the heating product.

* Multi-factor login support

Add multi factor login support and device login.

* Update Hive API to use device login

Update the Hive API and API test script to use the loginDevice. Update FHEM HiveHome object to use the updated API. Using manually file stored version of the device credentials.

* Tidy up device state conditions

* Add AWS Cognito Idp authentication (#10)

* Initial check-in

Initial work at demonstrating that I can implement the AWS login process for HiveHome.

* Tidy up test

Remove unnecessary use libraries and comments.

* Initial AWSCognitoIdp module

* Refactor and added RefreshTokens

 - Created separate initAuthentication and challengeResponse methods and called these from the loginSRP method.
- Requested the Math::BigInt use the GMP library to improve performance.
- Moved SRP initialisation code from the constructor to the loginSRP method.
- Added logging 
- Created helper methods to generate the AWS URL and request headers.
- Added a refreshToken method
- Added AWS ConfirmDevice functionality to the loginSRP method as this is required to be able to refresh the tokens.

* Refactor

Moved all AWS functionality to a new class and modified this to use the class functionality.

* Tidy up of previous commits.

* Further refactoring

Remove duplication and create new methods for specific functions.

* Add AWSCognitoIdp to HiveHome

Replace the original method of logging in to beehive with AWS Cognito Idp. Add support for using the refresh tokens rather than logging in again each time the current token expires.
Update the test scripts to include the extra parameter used by AWS Cognito Idp to support the refresh token.

* Initial work

Created a new library to hold common HiveHome functions. Moved some of the existing common functions to the new library.
Modified how the heating schedule and target are displayed so that they do not show the temperature offset. The temperature offset is applied to the schedule before it is sent to the device.
Modified how the target is used within the library to ensure the offset is applied to it where required.

* c

c

* TRVs no longer have zone IDs

Remove the zone ids checks for TRVS, assume they all belong to the heating product.

* Multi-factor login support

Add multi factor login support and device login.

* Update Hive API to use device login

Update the Hive API and API test script to use the loginDevice. Update FHEM HiveHome object to use the updated API. Using manually file stored version of the device credentials.

* Tidy up device state conditions